requirements.txt
python-dotenv
requests
colorama
fastapi
uvicorn
pydantic